拖拽红点到一定距离可清空购物车

+++ app/src/main/java/uk/co/ribot/androidboilerplate/ui/activity/PlaceOrderProductListImproveActivity.java
@@ -174,7 +174,7 @@ public class PlaceOrderProductListImproveActivity extends BaseActivity implement
             if (count == null) {
                 return 0;
             }
-            return count;
+            return count.doubleValue();
         }
 
     };
